feat(console): move status pills to a strip above the composer
Flip the wiring so the seven readiness pills plus the task-10 retrieval
Scope chip render in a new ConsoleStatusChips strip directly above the
composer, and slim ConsoleControlBar to the action row only (height 2 -> 1).
The Scope-chip machinery (ConsoleScopeChip, _scope_chip_render, _scope_chip,
sync_scope_chip) and the approvals-review action move to ConsoleStatusChips;
chat_screen composes the strip, syncs it alongside the control bar, and
retargets the scope-chip sync + ConsoleScopeChip import to the new module.
