+++ tldw_chatbook/css/tldw_cli_modular.tcss
@@ -1,7 +1,7 @@
 /* ========================================
  * GENERATED FILE - DO NOT EDIT DIRECTLY
  * ======================================== 
- * Generated: 2026-07-21 22:46:40
+ * Generated: 2026-07-21 22:49:12
  * 
  * This file is automatically generated by build_css.py
  * Edit the individual module files in core/, layout/, 
@@ -3429,6 +3429,16 @@ ConsoleSystemPromptModal {
     max-height: 1;
     layout: horizontal;
     align: left middle;
+}
+
+#console-status-chips {
+    width: 100%;
+    min-width: 0;
+    height: 1;
+    min-height: 1;
+    layout: horizontal;
+    border: none;
+    padding: 0 1;
 }
 
 .console-control-chip {
